Revert "#minor refacto and support scaleway api"
